- Error en la visualització de tasques amb data limit - Exportació de dades de l'entorn - Enviament de correus de documents signats i/o registrats
git-svn-id: svn://svn.code.sf.net/p/gehelium/code/helium-core/branches/branch-2.0@96 e3271b2b-fd41-4d56-bedf-4b5636213925

+++ presentacio-src/net/conselldemallorca/helium/presentacio/mvc/ExpedientDocumentEsborrarController.java
@@ -58,11 +58,13 @@ public class ExpedientDocumentEsborrarController extends BaseController {
 			if (potModificarExpedient(expedient)) {
 				try {
 					DocumentDto doc = expedientService.getDocument(docId, false, false);
-					if (!doc.isSignat()) {
+					if (!doc.isSignat() && !doc.isRegistrat()) {
 						expedientService.deleteDocument(id, docId);
 						missatgeInfo(request, "El document ha estat esborrat del procés");
-					} else {
+					} else if (doc.isSignat()) {
 						missatgeError(request, "No es pot esborrar un document signat");
+					} else if (doc.isRegistrat()) {
+						missatgeError(request, "No es pot esborrar un document registrat");
 					}
 				} catch (Exception ex) {
 					missatgeError(request, "No s'ha pogut esborrar el document del procés", ex.getLocalizedMessage());
